+++ src/basic/pages/CartPage.tsx
@@ -1,4 +1,4 @@
-import { CartItem, Coupon } from "../../types";
+import { Coupon, type CartItem as CartItemType } from "../../types";
 import { Product } from "../features/product/types";
 import { ProductList } from "../features/product/components/ProductList";
 import {
@@ -11,6 +11,7 @@ import {
   getRemainingStock,
 } from "../features/product/libs";
 import { useCallback } from "react";
+import { CartItem } from "../features/cart/components/CartItem";
 
 interface ProductWithUI extends Product {
   description?: string;
@@ -21,8 +22,8 @@ interface CartPageProps {
   products: ProductWithUI[];
   filteredProducts: ProductWithUI[];
   searchValue: string;
-  cart: CartItem[];
-  setCart: (cart: CartItem[]) => void;
+  cart: CartItemType[];
+  setCart: (cart: CartItemType[]) => void;
   coupons: Coupon[];
   selectedCoupon: Coupon | null;
   setSelectedCoupon: (coupon: Coupon | null) => void;
@@ -57,7 +58,7 @@ export function CartPage({
     return getRemainingStock(product, cartQuantity);
   };
 
-  const getMaxApplicableDiscount = (item: CartItem): number => {
+  const getMaxApplicableDiscount = (item: CartItemType): number => {
     const { discounts } = item.product;
     const { quantity } = item;
 
@@ -75,7 +76,7 @@ export function CartPage({
     return baseDiscount;
   };
 
-  const calculateItemTotal = (item: CartItem): number => {
+  const calculateItemTotal = (item: CartItemType): number => {
     const { price } = item.product;
     const { quantity } = item;
     const discount = getMaxApplicableDiscount(item);
@@ -271,78 +272,15 @@ export function CartPage({
               </div>
             ) : (
               <div className="space-y-3">
-                {cart.map((item) => {
-                  const itemTotal = calculateItemTotal(item);
-                  const originalPrice = item.product.price * item.quantity;
-                  const hasDiscount = itemTotal < originalPrice;
-                  const discountRate = hasDiscount
-                    ? Math.round((1 - itemTotal / originalPrice) * 100)
-                    : 0;
-
-                  return (
-                    <div
-                      key={item.product.id}
-                      className="border-b pb-3 last:border-b-0"
-                    >
-                      <div className="flex justify-between items-start mb-2">
-                        <h4 className="text-sm font-medium text-gray-900 flex-1">
-                          {item.product.name}
-                        </h4>
-                        <button
-                          onClick={() => removeFromCart(item.product.id)}
-                          className="text-gray-400 hover:text-red-500 ml-2"
-                        >
-                          <svg
-                            className="w-4 h-4"
-                            fill="none"
-                            stroke="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                              strokeWidth={2}
-                              d="M6 18L18 6M6 6l12 12"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                      <div className="flex items-center justify-between">
-                        <div className="flex items-center">
-                          <button
-                            onClick={() =>
-                              updateQuantity(item.product.id, item.quantity - 1)
-                            }
-                            className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
-                          >
-                            <span className="text-xs">−</span>
-                          </button>
-                          <span className="mx-3 text-sm font-medium w-8 text-center">
-                            {item.quantity}
-                          </span>
-                          <button
-                            onClick={() =>
-                              updateQuantity(item.product.id, item.quantity + 1)
-                            }
-                            className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
-                          >
-                            <span className="text-xs">+</span>
-                          </button>
-                        </div>
-                        <div className="text-right">
-                          {hasDiscount && (
-                            <span className="text-xs text-red-500 font-medium block">
-                              -{discountRate}%
-                            </span>
-                          )}
-                          <p className="text-sm font-medium text-gray-900">
-                            {Math.round(itemTotal).toLocaleString()}원
-                          </p>
-                        </div>
-                      </div>
-                    </div>
-                  );
-                })}
+                {cart.map((item) => (
+                  <CartItem
+                    key={item.product.id}
+                    item={item}
+                    calculateItemTotal={calculateItemTotal}
+                    removeFromCart={removeFromCart}
+                    updateQuantity={updateQuantity}
+                  />
+                ))}
               </div>
             )}
           </section>
